requirements.txt
-e ../airefinery-sdk
python-dotenv>=1.0.0
rich>=13.0.0
